animations update
