+++ src/pages/cases/[id]/ui/case-main-info.component.tsx
@@ -10,7 +10,8 @@ import { BitGraduateAuthority } from "./custom-main/bit-graduate-authority.compo
 import { useMounted } from "@shared/lib/hooks/useMounted";
 import clsx from "clsx";
 import { VideoComponent } from "@shared/ui/video/video.component";
-import { FairyTaleButton } from "./fairy-tale-button/fairy-tale-button";
+import { FairyTaleButton } from "../../../../widgets/landing/generate-book/ui/fairy-tale-button/fairytale.button";
+import Link from "next/link";
 
 export interface ICaseInfoItemProps {
   data?: ICase["main"][0];
@@ -80,11 +81,7 @@ function CaseInfoItem({ data }: ICaseInfoItemProps) {
   return (
     <section className={clsx("pb-[7.5rem] t-xs:pb-16 last:pb-0", data?.sectionClassnames)}>
       <div className="container">
-        {id === "fairy-tale-ai" && (
-          <div className="fixed right-4 bottom-10 z-999">
-            <FairyTaleButton />
-          </div>
-        )}
+        {id === "fairy-tale-ai" && <FairyTaleButton />}
         <article className="max-w-[52.5rem]">
           <h2 className="font-bold text-[2.875rem] whitespace-pre-line t-xs:text-2xl">
             {data?.title[locale || "ru"]}
